+++ vscode-ng-language-service/server/src/utils.ts
@@ -261,3 +261,38 @@ function getMappedContextSpan(
     ? {start: contextSpanStart.pos, length: contextSpanEnd.pos - contextSpanStart.pos}
     : undefined;
 }
+
+export function getTokenAtPosition(sourceFile: ts.SourceFile, position: number): ts.Node {
+  let current: ts.Node = sourceFile;
+  while (true) {
+    const child = current
+      .getChildren(sourceFile)
+      .find((c) => c.getStart(sourceFile) <= position && c.getEnd() > position);
+    if (!child || child.kind === ts.SyntaxKind.EndOfFileToken) {
+      return current;
+    }
+    current = child;
+  }
+}
+
+export enum LanguageId {
+  TS = 'typescript',
+  HTML = 'html',
+}
+
+export function isAngularCore(path: string): boolean {
+  return isExternalAngularCore(path) || isInternalAngularCore(path);
+}
+
+export function isExternalAngularCore(path: string): boolean {
+  return /@angular\/core\/.+\.d\.ts$/.test(path);
+}
+
+export function isInternalAngularCore(path: string): boolean {
+  // path in g3
+  return (
+    path.endsWith('angular2/rc/packages/core/index.d.ts') ||
+    // angular/angular repository direct sources
+    path.includes('angular/packages/core/src')
+  );
+}
